requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
scikit-learn>=1.3.0
xgboost>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
python-dotenv>=1.0.0
alpaca-trade-api>=3.0.0
schedule>=1.2.0
